Integration into the project data structure vector and subsequent writing unit-tests for them on Google Test
