+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(list)
 
 set(CMAKE_CXX_STANDARD 17)
 
-add_executable(list main.cpp)
+add_executable(list main.cpp src/data/list.cpp src/data/list.h src/data/vector.cpp src/data/vector.h)
