+++ src/data/vector.cpp
@@ -1,0 +1,5 @@
+//
+// Created by EviiiiiL on 22.01.2023.
+//
+
+#include "vector.h"
