+++ src/data/vector.h
@@ -1,0 +1,8 @@
+//
+// Created by EviiiiiL on 22.01.2023.
+//
+
+#ifndef LIST_VECTOR_H
+#define LIST_VECTOR_H
+
+#endif //LIST_VECTOR_H
